Skip database related tests if no MySQL database is available

+++ lib/mwlib/tests/MW/Cache/DBTest.php
@@ -25,7 +25,7 @@ class DBTest extends \PHPUnit_Framework_TestCase
 	 */
 	protected function setUp()
 	{
-		if( \TestHelperMw::getConfig()->get( 'resource/db/adapter', false ) === false ) {
+		if( \TestHelperMw::getConfig()->get( 'resource/db/adapter', false ) !== 'mysql' ) {
 			$this->markTestSkipped( 'No database configured' );
 		}
 
@@ -125,13 +125,16 @@ class DBTest extends \PHPUnit_Framework_TestCase
 	 */
 	protected function tearDown()
 	{
-		$this->dbm = \TestHelperMw::getDBManager();
-		$conn = $this->dbm->acquire();
+		if( \TestHelperMw::getConfig()->get( 'resource/db/adapter', false ) === 'mysql' )
+		{
+			$this->dbm = \TestHelperMw::getDBManager();
+			$conn = $this->dbm->acquire();
 
-		$conn->create( 'DROP TABLE "mw_cache_tag_test"' )->execute()->finish();
-		$conn->create( 'DROP TABLE "mw_cache_test"' )->execute()->finish();
+			$conn->create( 'DROP TABLE "mw_cache_tag_test"' )->execute()->finish();
+			$conn->create( 'DROP TABLE "mw_cache_test"' )->execute()->finish();
 
-		$this->dbm->release( $conn );
+			$this->dbm->release( $conn );
+		}
 	}
